app/requirements.txt
numpy
Pillow
prometheus-fastapi-instrumentator
psutil
prometheus_client
uvicorn
fastapi